+++ amqp-kit/publisher.go
@@ -1,11 +1,18 @@
 package amqp_kit
 
 import (
+	"context"
+	"log"
+
+	"github.com/opentracing-contrib/go-amqp/amqptracer"
+	"github.com/opentracing/opentracing-go"
+	"github.com/opentracing/opentracing-go/ext"
 	"github.com/streadway/amqp"
 )
 
 type Publisher struct {
-	ch Channel
+	ch          Channel
+	ServiceName string
 }
 
 // NewPublisher constructs a usable Publisher for a single remote method.
@@ -27,4 +34,30 @@ func (p Publisher) Publish(exchange, key, corID string, body []byte) (err error)
 	}
 
 	return err
+}
+
+//publish message to AMQP with tracing and span context
+func (p *Publisher) PublishWithTracing(ctx context.Context, exchange, key, corID string, body []byte) (err error) {
+	span, _ := opentracing.StartSpanFromContext(ctx, `publish_key: `+key)
+	defer span.Finish()
+
+	ext.SpanKind.Set(span, ext.SpanKindProducerEnum)
+	span.SetTag("key", key)
+	span.SetTag("exchange", exchange)
+	span.SetTag("corID", corID)
+
+	msg := amqp.Publishing{
+		Headers:       amqp.Table{},
+		ContentType:   "application/json",
+		CorrelationId: corID,
+		Body:          body,
+		DeliveryMode:  amqp.Persistent,
+	}
+
+	// Inject the span context into the AMQP header.
+	if err := amqptracer.Inject(span, msg.Headers); err != nil {
+		log.Printf("publish: error inject headers: %s", err)
+	}
+
+	return p.ch.Publish(exchange, key, false, false, msg)
 }