+++ tracing/tracing_test.go
@@ -2,63 +2,78 @@ package tracing
 
 import (
 	"context"
-	"net/http"
-	"net/http/httptest"
-	"testing"
+	"fmt"
+	"strings"
 
-	opentracing "github.com/opentracing/opentracing-go"
-	"github.com/opentracing/opentracing-go/ext"
-	"github.com/opentracing/opentracing-go/mocktracer"
-	"github.com/stretchr/testify/suite"
+	"github.com/opentracing/opentracing-go"
+	"github.com/uber/jaeger-client-go"
+	"github.com/uber/jaeger-client-go/config"
 )
 
-const serviceName = "test"
+func (ts *testSuite) TestCustomTracer() {
+	obs := &testObserver{}
+	tracerRoot, err := CreateTracer("newService", "", config.ContribObserver(obs))
+	ts.Require().NoError(err)
+	opentracing.SetGlobalTracer(tracerRoot.Tracer)
+	defer tracerRoot.Close()
 
-type testSuite struct {
-	suite.Suite
-	testServer *httptest.Server
-	client     *Client
-}
-
-func (ts *testSuite) TestDoWithTracing() {
-	beforeSpan := opentracing.GlobalTracer().StartSpan("to_inject").(*mocktracer.MockSpan)
-	defer beforeSpan.Finish()
-
-	req, err := http.NewRequest("", ts.testServer.URL, nil)
-	ts.NoError(err)
-
+	beforeSpan := opentracing.GlobalTracer().StartSpan("to_pub_inject")
 	beforeCtx := opentracing.ContextWithSpan(context.Background(), beforeSpan)
-	_, err = ts.client.DoWithTracing(beforeCtx, req)
-	ts.NoError(err)
 
-	finishedSpans := opentracing.GlobalTracer().(*mocktracer.MockTracer).FinishedSpans()
-	ts.Len(finishedSpans, 1)
+	tracer, err := CreateTracer("newService", "", config.ContribObserver(obs))
+	ts.Require().NoError(err)
+	sp := tracer.CreateSpan(beforeCtx, "insert")
+	sp.Finish()
 
-	finishedSpan := finishedSpans[0]
-	ts.Equal(serviceName, finishedSpan.Tag(string(ext.PeerService)))
+	beforeSpan.Finish()
 
-	beforeContext := beforeSpan.Context().(mocktracer.MockSpanContext)
-	ts.Equal(beforeContext.SpanID, finishedSpan.ParentID)
+	finishedSpans := obs.spans
+	ts.Require().Len(finishedSpans, 2)
+	second := strings.Split(fmt.Sprintf("%s", finishedSpans[1]), ":")
+	first := strings.Split(fmt.Sprintf("%s", finishedSpans[0]), ":")
 
-	ts.testServer.Close()
-	_, err = ts.client.DoWithTracing(beforeCtx, req)
-	ts.Error(err)
-	finishedSpans = opentracing.GlobalTracer().(*mocktracer.MockTracer).FinishedSpans()
-	ts.Len(finishedSpans, 2)
-	ts.Len(finishedSpans[1].Logs(), 1)
+	ts.Equal(first[0], first[1])
+
+	ts.Equal(first[0], second[0])
+	ts.Equal(first[0], second[2])
+	ts.NotEqual(first[0], second[1])
 }
 
-func (ts *testSuite) SetupSuite() {
-	ts.client = &Client{ServiceName: serviceName}
-	ts.testServer = httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {}))
-
-	tracer := mocktracer.New()
-	opentracing.SetGlobalTracer(tracer)
+//observer from tests
+type testObserver struct {
+	spans []opentracing.Span
 }
 
-func (ts *testSuite) TearDownSuite() {
-	ts.testServer.Close()
+type testSpanObserver struct {
+	operationName string
+	tags          map[string]interface{}
+	finished      bool
+	span          opentracing.Span
 }
-func TestEventsSuite(t *testing.T) {
-	suite.Run(t, new(testSuite))
+
+func (o *testObserver) OnStartSpan(sp opentracing.Span, operationName string, options opentracing.StartSpanOptions) (jaeger.ContribSpanObserver, bool) {
+	tags := make(map[string]interface{})
+	for k, v := range options.Tags {
+		tags[k] = v
+	}
+
+	o.spans = append(o.spans, sp)
+
+	return &testSpanObserver{
+		operationName: operationName,
+		tags:          tags,
+		span:          sp,
+	}, true
+}
+
+func (o *testSpanObserver) OnSetOperationName(operationName string) {
+	o.operationName = operationName
+}
+
+func (o *testSpanObserver) OnSetTag(key string, value interface{}) {
+	o.tags[key] = value
+}
+
+func (o *testSpanObserver) OnFinish(options opentracing.FinishOptions) {
+	o.finished = true
 }